+++ backend/techpick-api/src/main/java/techpick/api/domain/pick/service/PickService.java
@@ -1,6 +1,5 @@
 package techpick.api.domain.pick.service;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
@@ -11,14 +10,12 @@ import org.springframework.transaction.annotation.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import techpick.api.domain.folder.exception.ApiFolderException;
-import techpick.api.domain.link.dto.LinkInfo;
 import techpick.api.domain.pick.dto.PickCommand;
 import techpick.api.domain.pick.dto.PickMapper;
 import techpick.api.domain.pick.dto.PickResult;
 import techpick.api.domain.pick.exception.ApiPickException;
 import techpick.api.domain.tag.exception.ApiTagException;
 import techpick.api.infrastructure.folder.FolderDataHandler;
-import techpick.api.infrastructure.pick.PickBulkRepository;
 import techpick.api.infrastructure.pick.PickDataHandler;
 import techpick.api.infrastructure.tag.TagDataHandler;
 import techpick.core.model.folder.Folder;
@@ -34,7 +31,6 @@ public class PickService {
 	private final TagDataHandler tagDataHandler;
 	private final PickDataHandler pickDataHandler;
 	private final PickMapper pickMapper;
-	private final PickBulkRepository pickBulkRepository;
 	private final FolderDataHandler folderDataHandler;
 
 	@Transactional(readOnly = true)
@@ -127,18 +123,6 @@ public class PickService {
 		}
 
 		pickDataHandler.deletePickList(command);
-	}
-
-	@Transactional
-	public void saveBulkPick(Long userId, Long parentFolderId) {
-		List<PickCommand.Create> pickList = new ArrayList<>();
-		for (int i = 0; i < 10000; i++) {
-			LinkInfo linkInfo = new LinkInfo(String.valueOf(i), "링크 제목", "링크 설명", "링크 이미지 url", null);
-			PickCommand.Create command = new PickCommand.Create(userId, "테스트 제목", "테스트 메모", new ArrayList<>(),
-				parentFolderId, linkInfo);
-			pickList.add(command);
-		}
-		pickBulkRepository.bulkInsertPick(pickList);
 	}
 
 	/**
